refactor: 아키텍처 리팩토링 및 추상화
