+++ documentify-core/src/main/kotlin/io/github/bgmsound/documentify/core/Documentify.kt
@@ -1,11 +1,8 @@
 package io.github.bgmsound.documentify.core
 
+import io.github.bgmsound.documentify.core.documentation.specification.document.DocumentSpec
 import io.github.bgmsound.documentify.core.environment.StandaloneContext
-import io.github.bgmsound.documentify.core.specification.DocumentSpec
-import io.restassured.module.mockmvc.RestAssuredMockMvc.given
-import io.restassured.module.mockmvc.specification.MockMvcRequestSpecification
 import io.github.bgmsound.documentify.core.environment.StandaloneContext.Companion.controllers
-
 import org.junit.jupiter.api.extension.ExtendWith
 import org.springframework.restdocs.RestDocumentationContextProvider
 import org.springframework.restdocs.RestDocumentationExtension
@@ -18,24 +15,24 @@ import org.springframework.web.method.support.HandlerMethodArgumentResolver
 
 @ExtendWith(RestDocumentationExtension::class)
 abstract class Documentify {
-    private lateinit var spec: MockMvcRequestSpecification
+    private lateinit var mockMvc: MockMvc
 
     fun documentation(
         name: String,
-        specCustomizer: DocumentSpec.() -> Unit
+        specCustomizer: DocumentSpec.() -> Unit,
     ) {
         val documentSpec = DocumentSpec(name).also { specCustomizer(it) }
-        val emitter = RestDocsEmitter(documentSpec)
-        emitter.emit(spec)
+        val emitter = EmitterFactory.emitterOf(documentSpec)
+        emitter.emit(mockMvc)
     }
 
-    fun mockMvc(mock: MockMvc) {
-        spec = given().mockMvc(mock)
+    fun mockMvc(mockMvc: MockMvc) {
+        this.mockMvc = mockMvc
     }
 
     fun standalone(
         provider: RestDocumentationContextProvider,
-        contextCustomizer: StandaloneContext.() -> Unit
+        contextCustomizer: StandaloneContext.() -> Unit,
     ) {
         val standaloneContext = controllers().also(contextCustomizer)
         standalone(provider, standaloneContext)
@@ -43,17 +40,17 @@ abstract class Documentify {
 
     fun standalone(
         provider: RestDocumentationContextProvider,
-        standaloneContext: StandaloneContext
+        standaloneContext: StandaloneContext,
     ) {
         val mockMvc = standaloneContext.build(provider)
-        spec = given().mockMvc(mockMvc)
+        this.mockMvc = mockMvc
     }
 
     fun standalone(
         provider: RestDocumentationContextProvider,
         controllers: List<Any>,
         controllerAdvices: List<Any>,
-        argumentResolvers: List<HandlerMethodArgumentResolver>
+        argumentResolvers: List<HandlerMethodArgumentResolver>,
     ) {
         val standaloneContext = controllers(controllers)
             .controllerAdvices(controllerAdvices)
@@ -63,12 +60,12 @@ abstract class Documentify {
 
     fun webApplicationContext(
         provider: RestDocumentationContextProvider,
-        context: WebApplicationContext
+        context: WebApplicationContext,
     ) {
         val mockMvc = MockMvcBuilders
             .webAppContextSetup(context)
             .apply<DefaultMockMvcBuilder>(documentationConfiguration(provider))
             .build()
-        spec = given().mockMvc(mockMvc)
+        this.mockMvc = mockMvc
     }
 }
