+++ documentify-core/src/main/kotlin/io/github/bgmsound/documentify/core/EmitterFactory.kt
@@ -1,0 +1,11 @@
+package io.github.bgmsound.documentify.core
+
+import io.github.bgmsound.documentify.core.documentation.specification.document.DocumentSpec
+import io.github.bgmsound.documentify.core.emitter.DocumentEmitter
+import io.github.bgmsound.documentify.core.emitter.RestAssuredDocumentEmitter
+
+object EmitterFactory {
+    fun emitterOf(documentSpec: DocumentSpec): DocumentEmitter {
+        return RestAssuredDocumentEmitter(documentSpec)
+    }
+}
