+++ documentify-core/src/main/kotlin/io/github/bgmsound/documentify/core/documentation/APISpec.kt
@@ -1,0 +1,9 @@
+package io.github.bgmsound.documentify.core.documentation
+
+import org.springframework.restdocs.snippet.Snippet
+
+interface APISpec {
+
+    fun build(): List<Snippet>
+
+}
