+++ documentify-core/src/main/kotlin/io/github/bgmsound/documentify/core/documentation/specification/response/ResponseBodySpec.kt
@@ -1,0 +1,15 @@
+package io.github.bgmsound.documentify.core.documentation.specification.response
+
+import io.github.bgmsound.documentify.core.documentation.specification.BodySpec
+import org.springframework.restdocs.payload.PayloadDocumentation
+import org.springframework.restdocs.snippet.Snippet
+
+class ResponseBodySpec : BodySpec() {
+    override fun build(): List<Snippet> {
+        if (fields.isEmpty()) {
+            return emptyList()
+        }
+        val fieldsSnippet = PayloadDocumentation.responseFields(buildList { fields.forEach { addAll(it.build()) } })
+        return listOf(fieldsSnippet)
+    }
+}
