+++ documentify-core/src/main/kotlin/io/github/bgmsound/documentify/core/documentation/specification/response/ResponseHeaderSpec.kt
@@ -1,0 +1,15 @@
+package io.github.bgmsound.documentify.core.documentation.specification.response
+
+import io.github.bgmsound.documentify.core.documentation.specification.HeaderSpec
+import org.springframework.restdocs.headers.HeaderDocumentation
+import org.springframework.restdocs.snippet.Snippet
+
+class ResponseHeaderSpec : HeaderSpec() {
+    override fun build(): List<Snippet> {
+        if (headers.isEmpty()) {
+            return emptyList()
+        }
+        val headersSnippet = HeaderDocumentation.responseHeaders(headers.map { it.build() })
+        return listOf(headersSnippet)
+    }
+}
