+++ documentify-core/src/main/kotlin/io/github/bgmsound/documentify/core/emitter/DocumentEmitter.kt
@@ -1,0 +1,9 @@
+package io.github.bgmsound.documentify.core.emitter
+
+import org.springframework.test.web.servlet.MockMvc
+
+interface DocumentEmitter {
+
+    fun emit(mockMvc: MockMvc)
+
+}
