--- documentify-core/src/main/kotlin/io/github/bgmsound/documentify/core/specification/APISpec.kt
@@ -1,9 +1,0 @@
-package io.github.bgmsound.documentify.core.specification
-
-import org.springframework.restdocs.snippet.Snippet
-
-interface APISpec {
-
-    fun build(): List<Snippet>
-
-}
