--- documentify-core/src/main/kotlin/io/github/bgmsound/documentify/core/specification/HttpSpec.kt
@@ -1,10 +1,0 @@
-package io.github.bgmsound.documentify.core.specification
-
-abstract class HttpSpec(
-    private val header: HeaderSpec,
-    private val body: BodySpec
-) : APISpec {
-    val headers get() = header.headers()
-    val fields get() = body.fields()
-    val schema get() = body.schema()
-}
