--- documentify-core/src/main/kotlin/io/github/bgmsound/documentify/core/specification/element/Parameter.kt
@@ -1,11 +1,0 @@
-package io.github.bgmsound.documentify.core.specification.element
-
-import org.springframework.restdocs.request.ParameterDescriptor
-
-open class Parameter(
-    private val descriptor: ParameterDescriptor
-) : SpecElement(descriptor) {
-    override val key: String get() = descriptor.name
-
-    fun build() = descriptor
-}
